+++ test/src/extractor/test_looper_layer.cpp
@@ -12,7 +12,7 @@ namespace
 {
 struct dummy_extractor
 {
-    stub::function<uint64_t()> timestamp;
+    stub::function<uint64_t()> decoding_timestamp;
     stub::function<void()> advance;
     stub::function<void()> reset;
     stub::function<bool()> at_end;
@@ -26,13 +26,13 @@ TEST(extractor_test_looper_layer, not_looping)
     extractor_stack extractor;
     dummy_extractor& layer = extractor;
     extractor.disable_looping();
-    layer.timestamp.set_return(1);
+    layer.decoding_timestamp.set_return(1);
     layer.at_end.set_return(false, false, true);
     extractor.advance();
     extractor.advance();
     extractor.advance();
     EXPECT_EQ(0U, layer.reset.calls());
-    EXPECT_EQ(1U, extractor.timestamp());
+    EXPECT_EQ(1U, extractor.decoding_timestamp());
     EXPECT_EQ(3U, layer.advance.calls());
 }
 
@@ -40,21 +40,21 @@ TEST(extractor_test_looper_layer, looping)
 {
     extractor_stack extractor;
     dummy_extractor& layer = extractor;
-    layer.timestamp.set_return(1);
+    layer.decoding_timestamp.set_return(1);
     layer.at_end.set_return(false, false, true, false, false, true);
     EXPECT_EQ(0U, extractor.loops());
     extractor.advance();
     extractor.advance();
     extractor.advance();
     EXPECT_EQ(1U, layer.reset.calls());
-    EXPECT_EQ(2U, extractor.timestamp());
+    EXPECT_EQ(2U, extractor.decoding_timestamp());
     EXPECT_EQ(3U, layer.advance.calls());
     EXPECT_EQ(1U, extractor.loops());
     extractor.advance();
     extractor.advance();
     extractor.advance();
     EXPECT_EQ(2U, layer.reset.calls());
-    EXPECT_EQ(3U, extractor.timestamp());
+    EXPECT_EQ(3U, extractor.decoding_timestamp());
     EXPECT_EQ(6U, layer.advance.calls());
     EXPECT_EQ(2U, extractor.loops());
 }